requirements.txt
-i https://pypi.python.org/simple
amqp==5.1.1 ; python_version >= '3.6'
app-common-python==0.2.5
asgiref==3.6.0 ; python_version >= '3.7'
billiard==3.6.4.0
boto3==1.24.24
botocore==1.27.96 ; python_version >= '3.7'
celery==5.2.2
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
click-didyoumean==0.3.0 ; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.6
django==4.1.7
django-cors-headers==3.13.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==22.1
django-prometheus==2.2.0
django-tenants==3.4.8
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
ecs-logging==2.0.0
filelock==3.12.0 ; python_version >= '3.7'
gunicorn==20.1.0
idna==3.4 ; python_version >= '3.5'
jinja2==2.11.3
jmespath==1.0.1 ; python_version >= '3.7'
kafka-python==2.0.2
kombu==5.3.0b3 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
packaging==23.1 ; python_version >= '3.7'
platformdirs==3.5.0 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
prometheus-client==0.15.0
prompt-toolkit==3.0.38 ; python_full_version >= '3.7.0'
psycopg2==2.9.5
psycopg2-binary==2.9.5
py==1.11.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
python-dateutil==2.8.2
pytz==2022.2.1
redis==3.5.3
requests==2.28.2
s3transfer==0.6.1 ; python_version >= '3.7'
sentry-sdk==1.18.0
setuptools==67.7.2 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.4.3
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
tox==3.23.0
typing-extensions==4.5.0 ; python_version < '3.10'
tzdata==2022.2
unicodecsv==0.14.1
urllib3==1.26.15
vine==5.0.0 ; python_version >= '3.6'
virtualenv==20.23.0 ; python_version >= '3.7'
watchtower==3.0.0
wcwidth==0.2.6
whitenoise==6.4.0

django-treebeard~=4.7
Faker~=18.4.0
click~=8.1.3
PyYAML~=6.0
python-dotenv~=0.21.1
websockets~=10.4
pip~=23.1.2
attrs~=22.2.0
wheel~=0.38.1
cryptography~=39.0.2
docutils~=0.17.1
Sphinx~=4.3.2
pyasn1~=0.4.8
idna~=3.4
setuptools~=67.7.2
six~=1.16.0
certifi~=2023.5.7
cffi~=1.15.1